// src/app/core/services/market/types/market.ts
export class MarketType {
    name: string;
    totalAmount: number;
    minStake: string;
    maxStake: string;
    maxProfit: string;
    time: string;
    overlayMessage: string;
    isOverlayMessageVisible: boolean;
    isInPlay: boolean;
    isBetAllowed: boolean;
    sport: string;
    mt: number;
    iStatus: number;
    isShowCashout: boolean;
    }
    