// src/app/core/services/market/types/market-detail.ts
export class MarketDetail {
    betId: number;
    betDetailId: any;
    hdnProfitLoss: number;
    betTitle: string;
    lblProfitLoss: string;
    profitLoss: number;
    liability: any;
    lblLiability: string;
    backRate1: string;
    backRate2: string;
    backRate3: string;
    totalBackAmount1: string;
    totalBackAmount2: string;
    totalBackAmount3: string;
    layRate1: string;
    layRate2: string;
    layRate3: string;
    totalLayAmount1: string;
    totalLayAmount2: string;
    totalLayAmount3: string;
    isOverLayer: boolean;
    position: any;
    marketSuspend:boolean;
    cn : string;
    sd : number;
    cfn : string;
    jn : string;
    appRemovalDate: string;
    af : number;
    isJodiRates: boolean;
    tm: any;
    }
    